add select & delete cart data

+++ src/features/user-cart/actions.ts
@@ -1,16 +1,16 @@
 'use server'
 
-import { createUserCartDb, getCartDb, updateCartDb } from '@/db/utils/user-carts'
+import { createUserCartDb, deleteUserCartDb, getCartDb, updateUserCartDb } from '@/db/utils/user-carts'
 import { createAuthedServerAction } from '@/lib/server-actions/authed-server-action'
 import { CustomError } from '@/lib/server-actions/server-action'
 import { generateSecureRandomString } from '@/lib/utils'
-import { addToCartSchema, updateCartQtySchema } from '@/types/zod-schema'
-import { getCachedProduct } from '../products/lib/product-data'
+import { addToCartSchema, deleteCartSchema, updateCartQtySchema } from './lib/schema'
+import { getCachedProduct, getCachedProductStock } from '../products/lib/product-data'
 import { deleteCachedUserCart, deleteCachedUserCartProducts, getCachedUserCartProductIds } from './lib/cart-data'
 
 export const addToCartAction = createAuthedServerAction(addToCartSchema)
   .ratelimit({ key: 'add-to-cart', capacity: 10, refillRate: 10, refillPerSeconds: 10 })
-  .handle(async (session, { productId, quantity }) => {
+  .handle(async (session, { productId }) => {
     const id = generateSecureRandomString()
 
     const product = await getCachedProduct(productId)
@@ -18,9 +18,9 @@ export const addToCartAction = createAuthedServerAction(addToCartSchema)
 
     const cartProductIds = await getCachedUserCartProductIds(session.user.id)
     if (cartProductIds.includes(product.id)) throw new CustomError('forbidden', 'Duplicate items, Please reload the page')
+    const price = product.price.toString()
 
-    const price = (Number(product.price) * quantity).toString()
-    await createUserCartDb({ id, userId: session.user.id, productId, quantity, price })
+    await createUserCartDb({ id, userId: session.user.id, productId, quantity: 1, price })
     await deleteCachedUserCart(session.user.id)
   })
 
@@ -29,6 +29,17 @@ export const updateCartQtyAction = createAuthedServerAction(updateCartQtySchema)
   .handle(async (session, { cartId, quantity }) => {
     const cartData = await getCartDb(cartId)
     if (!cartData || cartData.userId !== session.user.id) throw new CustomError('not_found', 'Product data not found!')
-    await updateCartDb(cartId, { quantity })
+
+    const availableStock = await getCachedProductStock(cartData.productId)
+    if (quantity <= 0 || quantity > availableStock) throw new CustomError('not_found', 'Invalid product quantity')
+
+    await updateUserCartDb(session.user.id, cartId, { quantity })
     await deleteCachedUserCartProducts(session.user.id)
+  })
+
+export const deleteCartAction = createAuthedServerAction(deleteCartSchema)
+  .ratelimit({ key: 'delete-cart', capacity: 10, refillRate: 10, refillPerSeconds: 10 })
+  .handle(async (session, { cartIds }) => {
+    await deleteUserCartDb(session.user.id, cartIds)
+    await deleteCachedUserCart(session.user.id)
   })
